+++ crates/wdl-grammar/src/tree.rs
@@ -116,11 +116,11 @@ pub enum SyntaxKind {
     VersionKeyword,
     /// The `workflow` keyword token.
     WorkflowKeyword,
-    /// The reserved `Directory` type keyword token.
+    /// The 1.2 `Directory` type keyword token.
     DirectoryTypeKeyword,
-    /// The reserved `hints` keyword token.
+    /// The 1.2 `hints` keyword token.
     HintsKeyword,
-    /// The reserved `requirements` keyword token.
+    /// The 1.2 `requirements` keyword token.
     RequirementsKeyword,
     /// The `{` symbol token.
     OpenBrace,
@@ -214,6 +214,10 @@ pub enum SyntaxKind {
     OutputSectionNode,
     /// Represents a command section node.
     CommandSectionNode,
+    /// Represents a requirements section node.
+    RequirementsSectionNode,
+    /// Represents a requirements item node.
+    RequirementsItemNode,
     /// Represents a runtime section node.
     RuntimeSectionNode,
     /// Represents a runtime item node.